requirements.txt
fastapi
uvicorn
python-multipart
openai
pydantic
python-dotenv
pdf2image